+++ tests/test_dark_magic.py
@@ -1,4 +1,5 @@
 from __future__ import absolute_import, division, print_function
+import pickle
 
 import pytest
 from hypothesis import given
@@ -65,6 +66,11 @@ class SubSlots(SuperSlots):
 
 @attr.s(frozen=True, slots=True)
 class Frozen(object):
+    x = attr.ib()
+
+
+@attr.s(frozen=True, slots=False)
+class FrozenNoSlots(object):
     x = attr.ib()
 
 
@@ -177,3 +183,30 @@ class TestDarkMagic(object):
 
         assert e.value.args[0] == "can't set attribute"
         assert 1 == frozen.x
+
+    @pytest.mark.parametrize("cls",
+                             [C1, C1Slots, C2, C2Slots, Super, SuperSlots,
+                              Sub, SubSlots, Frozen, FrozenNoSlots])
+    @pytest.mark.parametrize("protocol",
+                             range(2, pickle.HIGHEST_PROTOCOL + 1))
+    def test_pickle_attributes(self, cls, protocol):
+        """
+        Pickling/un-pickling of Attribute instances works.
+        """
+        for attribute in attr.fields(cls):
+            assert attribute == pickle.loads(pickle.dumps(attribute, protocol))
+
+    @pytest.mark.parametrize("cls",
+                             [C1, C1Slots, C2, C2Slots, Super, SuperSlots,
+                              Sub, SubSlots, Frozen, FrozenNoSlots])
+    @pytest.mark.parametrize("protocol",
+                             range(2, pickle.HIGHEST_PROTOCOL + 1))
+    def test_pickle_object(self, cls, protocol):
+        """
+        Pickle object serialization works on all kinds of attrs classes.
+        """
+        if len(attr.fields(cls)) == 2:
+            obj = cls(123, 456)
+        else:
+            obj = cls(123)
+        assert repr(obj) == repr(pickle.loads(pickle.dumps(obj, protocol)))